Write BRAM content
